decorator visitor subclass
