+++ npe2/_inspection/__init__.py
@@ -1,0 +1,3 @@
+from ._visitors import NPE2PluginModuleVisitor, find_npe2_module_contributions
+
+__all__ = ["NPE2PluginModuleVisitor", "find_npe2_module_contributions"]
